+++ CHANGELOG.md
@@ -107,3 +107,8 @@
 - Fixed a bug where `echo` could crash on JavaScript if the module contains
   record variants with the same name as some built-in JavaScript objects.
   ([Louis Pilfold](https://github.com/lpil))
+
+- Fixed a bug where the compiler would highlight an entire double negation
+  expression as safe to remove, instead of just highlighting the double
+  negation.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))